+++ charts/camunda-platform-8.7/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [12.7.2](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-8.7-12.7.1...camunda-platform-8.7-12.7.2) (2025-11-28)
+
+
+### Dependencies
+
+* update camunda-platform-images (patch) ([#4792](https://github.com/camunda/camunda-platform-helm/issues/4792)) ([fd7294c](https://github.com/camunda/camunda-platform-helm/commit/fd7294c95d621b4d7d1c1d290b703d6209e61b44))
+* update patch-updates ([#4761](https://github.com/camunda/camunda-platform-helm/issues/4761)) ([89f5551](https://github.com/camunda/camunda-platform-helm/commit/89f55518ddeaeec8fb0423afd173cd39e631ea95))
+
+## [12.7.1](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-8.7-12.7.0...camunda-platform-8.7-12.7.1) (2025-11-25)
+
+
+### Dependencies
+
+* update camunda-platform-images (patch) ([#4763](https://github.com/camunda/camunda-platform-helm/issues/4763)) ([3fb6957](https://github.com/camunda/camunda-platform-helm/commit/3fb6957238e9daa4e7c6acdc9357dc6adeb2c0a9))
+* update camunda-platform-images (patch) ([#4777](https://github.com/camunda/camunda-platform-helm/issues/4777)) ([18de26f](https://github.com/camunda/camunda-platform-helm/commit/18de26fe264049929c2edddb8e9aa04f3c213b94))
+* update minor-updates (minor) ([#4712](https://github.com/camunda/camunda-platform-helm/issues/4712)) ([4cf435c](https://github.com/camunda/camunda-platform-helm/commit/4cf435c5aa989eaab1b0dde9cbc75fb694774854))
+* update minor-updates (minor) ([#4765](https://github.com/camunda/camunda-platform-helm/issues/4765)) ([54dc74d](https://github.com/camunda/camunda-platform-helm/commit/54dc74d5fed86702a26a63f247d7ccc25424946a))
+* update module golang.org/x/crypto to v0.45.0 [security] ([#4745](https://github.com/camunda/camunda-platform-helm/issues/4745)) ([1b31ade](https://github.com/camunda/camunda-platform-helm/commit/1b31aded5d1e7297e9648ad2e225b86f716a3b3e))
+* update patch-updates (patch) ([#4762](https://github.com/camunda/camunda-platform-helm/issues/4762)) ([f8e7bbd](https://github.com/camunda/camunda-platform-helm/commit/f8e7bbd242097bb2c7c7bfde54aa53b3a5077af2))
+
 ## [12.7.0](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-8.7-12.6.4...camunda-platform-8.7-12.7.0) (2025-11-20)